+++ src/schemathesis/runner/impl/core.py
@@ -14,22 +14,17 @@ from hypothesis.errors import InvalidArgument
 from hypothesis_jsonschema._canonicalise import HypothesisRefResolutionError
 from requests.auth import HTTPDigestAuth, _basic_auth_str
 
-from ...constants import (
-    DEFAULT_DEADLINE,
-    DEFAULT_STATEFUL_RECURSION_LIMIT,
-    RECURSIVE_REFERENCE_ERROR_MESSAGE,
-    USER_AGENT,
-    DataGenerationMethod,
-)
+from ...constants import DEFAULT_DEADLINE, RECURSIVE_REFERENCE_ERROR_MESSAGE, USER_AGENT, DataGenerationMethod
 from ...exceptions import CheckFailed, InvalidRegularExpression, InvalidSchema, NonCheckError, get_grouped_exception
 from ...hooks import HookContext, get_all_by_name
 from ...models import APIOperation, Case, Check, CheckFunction, Status, TestResult, TestResultSet
 from ...runner import events
 from ...schemas import BaseSchema
-from ...stateful import Feedback, Stateful
+from ...stateful import Stateful
 from ...targets import Target, TargetContext
 from ...types import RawAuth
 from ...utils import GenericResponse, Ok, WSGIResponse, capture_hypothesis_output, format_exception
+from ..events import Phase
 from ..serialization import SerializedTestResult
 
 
@@ -55,16 +50,25 @@ class BaseRunner:
     exit_first: bool = attr.ib(default=False)  # pragma: no mutate
     dry_run: bool = attr.ib(default=False)  # pragma: no mutate
     stateful: Optional[Stateful] = attr.ib(default=None)  # pragma: no mutate
-    stateful_recursion_limit: int = attr.ib(default=DEFAULT_STATEFUL_RECURSION_LIMIT)  # pragma: no mutate
 
     def execute(self) -> Generator[events.ExecutionEvent, None, None]:
-        """Common logic for all runners."""
+        """Common logic for all runners.
+
+        Runs all the available phases.
+        """
         results = TestResultSet()
 
         initialized = events.Initialized.from_schema(schema=self.schema)
         yield initialized
 
-        for event in self._execute(results):
+        yield from self.run_unit_tests(results)
+        yield from self.run_stateful_tests(results)
+
+        yield events.Finished.from_results(results=results, running_time=time.monotonic() - initialized.start_time)
+
+    def run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+        yield events.BeforePhase(phase=Phase.unit_testing)
+        for event in self._run_unit_tests(results):
             yield event
             if (
                 self.exit_first
@@ -72,58 +76,45 @@ class BaseRunner:
                 and event.status in (Status.error, Status.failure)
             ):
                 break
+            yield event
+        yield events.AfterPhase(phase=Phase.unit_testing)
 
-        yield events.Finished.from_results(results=results, running_time=time.monotonic() - initialized.start_time)
-
-    def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+    def _run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         raise NotImplementedError
+
+    def run_stateful_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+        yield events.BeforePhase(phase=Phase.stateful_testing)
+        yield events.AfterPhase(phase=Phase.stateful_testing)
 
     def _run_tests(
         self,
-        maker: Callable,
         template: Callable,
         settings: hypothesis.settings,
         seed: Optional[int],
         results: TestResultSet,
-        recursion_level: int = 0,
         **kwargs: Any,
     ) -> Generator[events.ExecutionEvent, None, None]:
         """Run tests and recursively run additional tests."""
-        if recursion_level > self.stateful_recursion_limit:
-            return
-        for result, data_generation_method in maker(template, settings, seed):
+        for result, data_generation_method in self.schema.get_all_tests(template, settings, seed):
             if isinstance(result, Ok):
                 operation, test = result.ok()
-                feedback = Feedback(self.stateful, operation)
                 for event in run_test(
                     operation,
                     test,
                     results=results,
-                    feedback=feedback,
-                    recursion_level=recursion_level,
                     data_generation_method=data_generation_method,
                     **kwargs,
                 ):
                     yield event
                     if isinstance(event, events.Interrupted):
                         return
-                # Additional tests, generated via the `feedback` instance
-                yield from self._run_tests(
-                    feedback.get_stateful_tests,
-                    template,
-                    settings,
-                    seed,
-                    recursion_level=recursion_level + 1,
-                    results=results,
-                    **kwargs,
-                )
             else:
                 # Schema errors
-                yield from handle_schema_error(result.err(), results, data_generation_method, recursion_level)
+                yield from handle_schema_error(result.err(), results, data_generation_method)
 
 
 def handle_schema_error(
-    error: InvalidSchema, results: TestResultSet, data_generation_method: DataGenerationMethod, recursion_level: int
+    error: InvalidSchema, results: TestResultSet, data_generation_method: DataGenerationMethod
 ) -> Generator[events.ExecutionEvent, None, None]:
     if error.method is not None:
         assert error.path is not None
@@ -136,9 +127,7 @@ def handle_schema_error(
         )
         result.add_error(error)
 
-        yield events.BeforeExecution(
-            method=method, path=error.full_path, relative_path=error.path, recursion_level=recursion_level
-        )
+        yield events.BeforeExecution(method=method, path=error.full_path, relative_path=error.path)
         yield events.AfterExecution(
             method=method,
             path=error.full_path,
@@ -163,7 +152,6 @@ def run_test(  # pylint: disable=too-many-locals
     targets: Iterable[Target],
     results: TestResultSet,
     headers: Optional[Dict[str, Any]],
-    recursion_level: int,
     **kwargs: Any,
 ) -> Generator[events.ExecutionEvent, None, None]:
     """A single test run with all error handling needed."""
@@ -173,7 +161,7 @@ def run_test(  # pylint: disable=too-many-locals
         overridden_headers=headers,
         data_generation_method=data_generation_method,
     )
-    yield events.BeforeExecution.from_operation(operation=operation, recursion_level=recursion_level)
+    yield events.BeforeExecution.from_operation(operation=operation)
     hypothesis_output: List[str] = []
     errors: List[Exception] = []
     test_start_time = time.monotonic()
@@ -375,7 +363,6 @@ def network_test(
     request_tls_verify: bool,
     store_interactions: bool,
     headers: Optional[Dict[str, Any]],
-    feedback: Feedback,
     max_response_time: Optional[int],
     dry_run: bool,
     errors: List[Exception],
@@ -396,25 +383,23 @@ def network_test(
                 timeout,
                 store_interactions,
                 headers,
-                feedback,
                 request_tls_verify,
                 max_response_time,
             )
-            add_cases(
-                case,
-                response,
-                _network_test,
-                checks,
-                targets,
-                result,
-                session,
-                timeout,
-                store_interactions,
-                headers,
-                feedback,
-                request_tls_verify,
-                max_response_time,
-            )
+    add_cases(
+        case,
+        response,
+        _network_test,
+        checks,
+        targets,
+        result,
+        session,
+        timeout,
+        store_interactions,
+        headers,
+        request_tls_verify,
+        max_response_time,
+    )
 
 
 def _network_test(
@@ -426,7 +411,6 @@ def _network_test(
     timeout: Optional[float],
     store_interactions: bool,
     headers: Optional[Dict[str, Any]],
-    feedback: Feedback,
     request_tls_verify: bool,
     max_response_time: Optional[int],
 ) -> requests.Response:
@@ -443,7 +427,6 @@ def _network_test(
     finally:
         if store_interactions:
             result.store_requests_response(response, status, check_results)
-    feedback.add_test_case(case, response)
     return response
 
 
@@ -472,7 +455,6 @@ def wsgi_test(
     auth_type: Optional[str],
     headers: Optional[Dict[str, Any]],
     store_interactions: bool,
-    feedback: Feedback,
     max_response_time: Optional[int],
     dry_run: bool,
     errors: List[Exception],
@@ -480,9 +462,7 @@ def wsgi_test(
     with ErrorCollector(errors):
         headers = _prepare_wsgi_headers(headers, auth, auth_type)
         if not dry_run:
-            response = _wsgi_test(
-                case, checks, targets, result, headers, store_interactions, feedback, max_response_time
-            )
+            response = _wsgi_test(case, checks, targets, result, headers, store_interactions, max_response_time)
             add_cases(
                 case,
                 response,
@@ -492,7 +472,6 @@ def wsgi_test(
                 result,
                 headers,
                 store_interactions,
-                feedback,
                 max_response_time,
             )
 
@@ -504,7 +483,6 @@ def _wsgi_test(
     result: TestResult,
     headers: Dict[str, Any],
     store_interactions: bool,
-    feedback: Feedback,
     max_response_time: Optional[int],
 ) -> WSGIResponse:
     with catching_logs(LogCaptureHandler(), level=logging.DEBUG) as recorded:
@@ -524,7 +502,6 @@ def _wsgi_test(
     finally:
         if store_interactions:
             result.store_wsgi_response(case, response, headers, elapsed, status, check_results)
-    feedback.add_test_case(case, response)
     return response
 
 
@@ -555,7 +532,6 @@ def asgi_test(
     result: TestResult,
     store_interactions: bool,
     headers: Optional[Dict[str, Any]],
-    feedback: Feedback,
     max_response_time: Optional[int],
     dry_run: bool,
     errors: List[Exception],
@@ -565,21 +541,18 @@ def asgi_test(
         headers = headers or {}
 
         if not dry_run:
-            response = _asgi_test(
-                case, checks, targets, result, store_interactions, headers, feedback, max_response_time
-            )
-            add_cases(
-                case,
-                response,
-                _asgi_test,
-                checks,
-                targets,
-                result,
-                store_interactions,
-                headers,
-                feedback,
-                max_response_time,
-            )
+            response = _asgi_test(case, checks, targets, result, store_interactions, headers, max_response_time)
+    add_cases(
+        case,
+        response,
+        _asgi_test,
+        checks,
+        targets,
+        result,
+        store_interactions,
+        headers,
+        max_response_time,
+    )
 
 
 def _asgi_test(
@@ -589,7 +562,6 @@ def _asgi_test(
     result: TestResult,
     store_interactions: bool,
     headers: Optional[Dict[str, Any]],
-    feedback: Feedback,
     max_response_time: Optional[int],
 ) -> requests.Response:
     response = case.call_asgi(headers=headers)
@@ -605,5 +577,4 @@ def _asgi_test(
     finally:
         if store_interactions:
             result.store_requests_response(response, status, check_results)
-    feedback.add_test_case(case, response)
     return response