+++ src/schemathesis/runner/impl/solo.py
@@ -15,11 +15,10 @@ class SingleThreadRunner(BaseRunner):
 
     request_tls_verify: Union[bool, str] = attr.ib(default=True)  # pragma: no mutate
 
-    def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+    def _run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         auth = get_requests_auth(self.auth, self.auth_type)
         with get_session(auth) as session:
             yield from self._run_tests(
-                self.schema.get_all_tests,
                 network_test,
                 self.hypothesis_settings,
                 self.seed,
@@ -38,9 +37,8 @@ class SingleThreadRunner(BaseRunner):
 
 @attr.s(slots=True)  # pragma: no mutate
 class SingleThreadWSGIRunner(SingleThreadRunner):
-    def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+    def _run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         yield from self._run_tests(
-            self.schema.get_all_tests,
             wsgi_test,
             self.hypothesis_settings,
             self.seed,
@@ -58,9 +56,8 @@ class SingleThreadWSGIRunner(SingleThreadRunner):
 
 @attr.s(slots=True)  # pragma: no mutate
 class SingleThreadASGIRunner(SingleThreadRunner):
-    def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
+    def _run_unit_tests(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         yield from self._run_tests(
-            self.schema.get_all_tests,
             asgi_test,
             self.hypothesis_settings,
             self.seed,