+++ test/test_stateful.py
@@ -4,18 +4,8 @@ import pytest
 
 import schemathesis
 from schemathesis.specs.openapi import expressions
-from schemathesis.stateful import ParsedData
-from schemathesis.utils import NOT_SET
 
 from .apps.utils import OpenAPIVersion
-
-
-@pytest.mark.parametrize(
-    "parameters, body", (({"a": 1}, None), ({"a": 1}, NOT_SET), ({"a": 1}, {"value": 1}), ({"a": 1}, [1, 2, 3]))
-)
-def test_hashable(parameters, body):
-    # All parsed data should be hashable
-    hash(ParsedData(parameters, body))
 
 
 @pytest.fixture